Move get/cancel order from actions to utils.
